Add TreeDiff trait to reuse tree functions
